Make Documentation More Clear about Misuse
Improve some of the documentation to make it more clear that the only output you can trust from AntiSamy is the cleanResults text. `errorMessages` & `GetNumberOfErrors` are just aides primarily for testing, but are not security controls (i.e., you can't rely on them to decide if the input is 'safe' or not). This was specified in the README but not as clearly here.